requirements.txt
asgiref==3.5.2
coverage==6.4.4
Django==4.1
flake8==5.0.4
isort==5.10.1
mccabe==0.7.0
Pillow==9.2.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==1.0.0
sqlparse==0.4.2
tzdata==2022.2
